+++ atomic_red_team/validator.py
